Extracting getConfig function
